Circular stats
Moved the instrument-independent circular statistics routines from the
omni_hro instrument file into the utilities file.  This required moving
the tests for these routines, as well.

Also fixed a bug in the keyword arguments used in two of the omni_hro
tests.
